crypto: move room_settings tests to their own file
